add is-github-fork-synced-with-upstream.sh

+++ zabbix/update-zabbix-metadata.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
+# https://stackoverflow.com/a/18216122
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
+
+cd /etc/zabbix/zabbix_agent2.d/bash_configs/zabbix
+
 cp zabbix_agent2.conf ../../../
 rm /etc/zabbix/zabbix_agent2.d/zabbix*.conf
 find -L /etc/zabbix/zabbix_agent2.d -name 'zabbix*.conf' | grep -v zabbix_agent2.conf | xargs -I{} cp {} /etc/zabbix/zabbix_agent2.d/
